pipeline/requirements.txt
python-dotenv
stomp.py
PyXB-X